pyffmpeg/requirements.txt
####### requirements.txt #######

###### Requirements without Version Specifiers ######
requests
datetime
ffmpeg-python

###### Requirements with Version Specifiers ######
# requests == 2.24.0             # Version Matching. Must be version 2.24.0

#
###### Use pip no cache ######
# pip install --no-cache-dir -r requirements.txt